HLS converte/requirements.txt
altgraph==0.17.2
bottle==0.12.19
bottle-websocket==0.2.9
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
Eel==0.14.0
future==0.18.2
gevent==21.12.0
gevent-websocket==0.10.1
greenlet==1.1.2
idna==3.3
mysql-connector-python==8.0.29
pefile==2021.9.3
protobuf==3.20.1
pycparser==2.21
pyinstaller==5.0.1
pyinstaller-hooks-contrib==2022.4
pyparsing==3.0.9
python-ffmpeg-video-streaming==0.1.15
pywin32-ctypes==0.2.0
requests==2.27.1
urllib3==1.26.9
whichcraft==0.6.1
zope.event==4.5.0
zope.interface==5.4.0
